Use i64 instead of uint64_t

+++ src/globals.h
@@ -21,7 +21,7 @@
 typedef unsigned char u8;
 typedef unsigned int u16;
 typedef unsigned long int u32;
-typedef uint64_t u64;
+typedef unsigned long long int u64;
 
 #define BYTE_TO_BINARY_PATTERN "%c%c%c%c%c%c%c%c"
 #define BYTE_TO_BINARY(byte)  \
